+++ src/Serialize.Linq/ExpressionContext.cs
@@ -16,8 +16,20 @@ namespace Serialize.Linq
 
         public ExpressionContext(IAssemblyLoader assemblyLoader)
         {
-            _assemblyLoader = assemblyLoader 
+            _assemblyLoader = assemblyLoader
                 ?? throw new ArgumentNullException(nameof(assemblyLoader));
+        }
+
+        public ExpressionContext(ITypeFilter typeFilter)
+            : this(new DefaultAssemblyLoader())
+        {
+            TypeFilter = typeFilter;
+        }
+
+        public ExpressionContext(IAssemblyLoader assemblyLoader, ITypeFilter typeFilter)
+            : this(assemblyLoader)
+        {
+            TypeFilter = typeFilter;
         }
 
         protected override IEnumerable<Assembly> GetAssemblies()